[VWLS-157] Mobile life input fixes
a. "choose lifestyle" - should be a dropdown. This is not working currently.
b. "choose lifestyle" text should be underlined
c. "choose lifestyle" text shouldn't be bold.
d. Same font style should be used for "enter device" and "choose lifestyle"
c. Mouse hover on the accessories (resulted products) is not working
d. When a letter I is typed - the typeahead text is shown in the device text box. It should not be the case. User should be selecting this.